requirement.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
blessed==1.20.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==5.0.3
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.7
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
sty==1.0.6
Twisted==24.3.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
